+++ src/types.ts
@@ -29,6 +29,7 @@ export interface Results {
 		bytes: Bytes
 		/** Note: Results use slashes as delimiters even on Windows. */
 		results: Record<FilePath, LanguageResult>
+		alternatives: Record<FilePath, LanguageResult[]>
 	}
 	languages: {
 		count: Integer