+++ test/unit.js
@@ -13,6 +13,7 @@ async function test([filename, fileContent = ''], [type, testVal]) {
 		'files': actual.files.results[filename],
 		'size': actual.files.bytes,
 		'count': actual.files.count,
+		'alternatives_count': Object.entries(actual.files.alternatives).length,
 	}[type];
 	const result = testContent === testVal;
 	i = `${+i + 1}`.padStart(2, '0');
@@ -36,6 +37,7 @@ async function unitTest() {
 	await test(['x.cpp'], ['files', 'C++']);
 	await test(['x.c'], ['files', 'C']);
 	await test(['x.R'], ['files', 'R']);
+	await test(['.m'], ['alternatives_count', 1])
 	desc('filenames');
 	await test(['Dockerfile'], ['files', 'Dockerfile']);
 	await test(['CMakeLists.txt'], ['files', 'CMake']);